requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
Django==4.0.5
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-tinymce==3.4.0
Pillow==9.1.1
pyenchant==3.2.2
sqlparse==0.4.2
